+++ tests/files/src/0674_unused_loop_variable.php
@@ -1,0 +1,7 @@
+<?php
+call_user_func (function () {
+    foreach ([1,2,3] as $i) {
+        $c = $i;
+        $c = $c + 1;  // should emit PhanUnusedVariable
+    }
+});