+++ NEWS.md
@@ -11,6 +11,7 @@ New features(Analysis):
 + Infer the return types of PHP 7.4's magic methods `__serialize()` and `__unserialize()`. (#2755)
   Improve analysis of return types of other magic methods such as `__sleep()`.
 + Support more of PHP 7.4's function signatures (e.g. `WeakReference`) (#2756)
++ Improve detection of unused variables inside of loops/branches.
 
 Plugins:
 + Detect some new php 7.3 functions (`array_key_first`, etc.) in `UseReturnValuePlugin`.